backend/requirements.txt
fastapi
pydantic
pydantic-settings
python-dotenv
google-generativeai
anthropic
openai
vnstock
passlib
python-jose
pandas
